refactor(api): move index and controller to evaluation context
