+++ api/src/evaluation/application/users/index.js
@@ -31,6 +31,31 @@ const register = async function (server) {
         tags: ['api', 'user'],
       },
     },
+    {
+      method: 'PATCH',
+      path: '/api/users/{id}/has-seen-new-dashboard-info',
+      config: {
+        pre: [
+          {
+            method: securityPreHandlers.checkRequestedUserIsAuthenticatedUser,
+            assign: 'requestedUserIsAuthenticatedUser',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.userId,
+          }),
+        },
+        handler: userController.rememberUserHasSeenNewDashboardInfo,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Sauvegarde le fait que l'utilisateur ait vu le message sur le nouveau dashboard" +
+            '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
+          "- Le contenu de la requête n'est pas pris en compte.",
+        ],
+        tags: ['api', 'user'],
+      },
+    },
   ]);
 };
