+++ api/src/evaluation/application/users/user-controller.js
@@ -1,3 +1,4 @@
+import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import * as userSerializer from '../../../shared/infrastructure/serializers/jsonapi/user-serializer.js';
 import { evaluationUsecases as usecases } from '../../domain/usecases/index.js';
 
@@ -8,6 +9,13 @@ const rememberUserHasSeenLevelSevenInfo = async function (request, h, dependenci
   return dependencies.userSerializer.serialize(updatedUser);
 };
 
-const userController = { rememberUserHasSeenLevelSevenInfo };
+const rememberUserHasSeenNewDashboardInfo = async function (request, h, dependencies = { userSerializer }) {
+  const authenticatedUserId = request.auth.credentials.userId;
+
+  const updatedUser = await libUsecases.rememberUserHasSeenNewDashboardInfo({ userId: authenticatedUserId });
+  return dependencies.userSerializer.serialize(updatedUser);
+};
+
+const userController = { rememberUserHasSeenLevelSevenInfo, rememberUserHasSeenNewDashboardInfo };
 
 export { userController };
